Refactor group and notice view models for ID handling
Changed group and notice view models to use String and Long types for group IDs instead of Int, improving consistency and type safety. Updated related repository and UI code to match new ID handling. Also simplified hashtag processing and fixed minor variable naming issues.

+++ app/src/main/java/com/example/studify/ui/noticeVM.kt
@@ -19,7 +19,7 @@ class noticeVM @Inject constructor(
     savedStateHandle: SavedStateHandle
 ) : ViewModel() {
 
-    val groupId = mutableStateOf(savedStateHandle.get<String>("groupid")!!.toInt())
+    val groupId = mutableStateOf(savedStateHandle.get<String>("groupid")!!.toLong())
     val notices = mutableStateListOf<AnnounceModel.AnnounceContent>()
     val query = mutableStateOf("")
     val selectedNotice = mutableStateOf<AnnounceModel.AnnounceContent?>(null)
